# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /acquisition-recommendations:
    get:
      operationId: listAcquisitionRecommendations
      tags: [acquisition]
      summary: List deterministic acquisition recommendations
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/RecommendationPriority"
        - name: blocked
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Owner-scoped recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcquisitionRecommendation"
    post:
      operationId: generateAcquisitionRecommendations
      tags: [acquisition]
      summary: Reconcile recommendations from current archive and provider evidence
      responses:
        "200":
          description: Current recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcquisitionRecommendation"
  /acquisition-recommendations/{id}:
    get:
      operationId: getAcquisitionRecommendation
      tags: [acquisition]
      summary: Inspect one acquisition recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Acquisition recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcquisitionRecommendation"
        "404":
          description: Recommendation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /review-items:
    get:
      operationId: listReviewItems
      tags: [review]
      summary: List and filter generic review items
      parameters:
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/ReviewItemState"
        - name: kind
          in: query
          schema:
            $ref: "#/components/schemas/ReviewItemKind"
      responses:
        "200":
          description: Review queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewItem"
    post:
      operationId: createReviewItem
      tags: [review]
      summary: Idempotently add a subject to the review queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewItem"
      responses:
        "201":
          description: Review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"
...
  /review-items/{id}:
    get:
      operationId: getReviewItem
      tags: [review]
      summary: Inspect a review item and decision history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"
        "404":
          description: Review item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /review-items/{id}/approve:
    post:
      operationId: approveReviewQueueItem
      tags: [review]
      summary: Approve a pending or reopened review item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionInput"
      responses:
        "200":
          description: Approved review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /review-items/{id}/reject:
    post:
      operationId: rejectReviewQueueItem
      tags: [review]
      summary: Reject a pending or reopened review item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionInput"
      responses:
        "200":
          description: Rejected review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"
  /review-items/{id}/defer:
    post:
      operationId: deferReviewQueueItem
      tags: [review]
      summary: Defer a pending or reopened review item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionInput"
      responses:
        "200":
          description: Deferred review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"
  /review-items/{id}/reopen:
    post:
      operationId: reopenReviewQueueItem
      tags: [review]
      summary: Reopen a decided review item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionInput"
      responses:
        "200":
          description: Reopened review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"
  /review-items/{id}/acquisition-job:
    post:
      operationId: createApprovedAcquisitionJob
      tags: [review, acquisition]
      summary: Explicitly create or return the acquisition job for an approved recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Recommendation and linked acquisition job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedAcquisitionResult"
        "400":
          description: Approval or provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /archive-operations:
    get:
      operationId: listArchiveOperations
      tags: [operations]
      summary: List owner-scoped approved archive operations
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ArchiveOperationStatus"
      responses:
        "200":
          description: Archive operations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchiveOperation"
    post:
      operationId: createArchiveOperation
      tags: [operations]
      summary: Plan an operation linked to an approved review item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArchiveOperation"
      responses:
        "201":
          description: Planned operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
        "400":
          description: Approval or input invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /archive-operations/{id}:
    get:
      operationId: getArchiveOperation
      tags: [operations]
      summary: Inspect one archive operation and its history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Archive operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
        "404":
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /archive-operations/{id}/preflight:
    post:
      operationId: preflightArchiveOperation
      tags: [operations]
      summary: Validate an approved operation without changing files
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Preflight result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
  /archive-operations/{id}/execute:
    post:
      operationId: executeArchiveOperation
      tags: [operations]
      summary: Explicitly execute a ready approved operation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationConfirmation"
      responses:
        "200":
          description: Operation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
        "400":
          description: Execution rejected or failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /archive-operations/{id}/cancel:
    post:
      operationId: cancelArchiveOperation
      tags: [operations]
      summary: Cancel an operation before execution
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Cancelled operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
  /archive-operations/{id}/retry:
    post:
      operationId: retryArchiveOperation
      tags: [operations]
      summary: Retry preflight for a failed or cancelled operation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Retried operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
  /archive-operations/{id}/rollback:
    post:
      operationId: rollbackArchiveOperation
      tags: [operations]
      summary: Explicitly roll back a completed operation where supported
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationConfirmation"
      responses:
        "200":
          description: Rolled back operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
...
  /action-proposals:
    get:
      operationId: listActionProposals
      tags: [actions]
      summary: List owner-scoped action proposals
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ActionProposalStatus"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/ActionType"
        - name: source
          in: query
          schema:
            $ref: "#/components/schemas/ActionSource"
      responses:
        "200":
          description: Action proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActionProposal"
  /action-proposals/{id}:
    get:
      operationId: getActionProposal
      tags: [actions]
      summary: Read a single action proposal with its steps and history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionProposal"
        "404":
          description: Action proposal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /action-proposals/{id}/selection:
    post:
      operationId: updateActionStepSelection
      tags: [actions]
      summary: Select or deselect individual steps before approval
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionStepSelection"
      responses:
        "200":
          description: Updated action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionProposal"
        "400":
          description: Selection rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /action-proposals/{id}/approve:
    post:
      operationId: approveActionProposal
      tags: [actions]
      summary: Record explicit durable approval for an action proposal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionApproval"
      responses:
        "200":
          description: Approved action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionProposal"
        "400":
          description: Approval rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /action-proposals/{id}/preflight:
    post:
      operationId: preflightActionProposal
      tags: [actions]
      summary: Re-check safety conditions without changing the archive
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      responses:
        "200":
          description: Preflighted action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionProposal"
        "400":
          description: Preflight rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /action-proposals/{id}/execute:
    post:
      operationId: executeActionProposal
      tags: [actions]
      summary: Execute an approved and preflighted action proposal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationConfirmation"
      responses:
        "200":
          description: Executed action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionProposal"
        "400":
          description: Execution rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /action-proposals/{id}/revert:
    post:
      operationId: revertActionProposal
      tags: [actions]
      summary: Restore the archive state changed by a completed proposal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationConfirmation"
      responses:
        "200":
          description: Reverted action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionProposal"
        "400":
          description: Revert rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /archive/naming-actions:
    get:
      operationId: getArchiveNamingActionCandidates
      tags: [actions, archive]
      summary: Preview the rename actions available for naming findings
      parameters:
        - name: confidence
          in: query
          schema: { type: string }
        - name: pattern
          in: query
          schema: { type: string }
        - name: mediaType
          in: query
          schema: { type: string }
        - name: volume
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: number, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Available naming actions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamingActionCandidates"
    post:
      operationId: planArchiveNamingNormalization
      tags: [actions, archive]
      summary: Create a reviewable rename proposal from naming findings
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanNamingNormalization"
      responses:
        "201":
          description: Created action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionProposal"
        "400":
          description: No actionable naming findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /acquisition-jobs/{id}/download:
    post:
      operationId: linkAcquisitionDownload
      tags: [acquisition]
      summary: Link an owner-scoped local download to an acquisition job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkAcquisitionDownload"
      responses:
        "200":
          description: Updated acquisition job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcquisitionJob"
  /acquisition-jobs/{id}/import:
    post:
      operationId: planApprovedAcquisitionImport
      tags: [acquisition, operations]
      summary: Plan an approved import from a completed verified download
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanAcquisitionImport"
      responses:
        "201":
          description: Planned import operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveOperation"
...
  /archive/reconciliation:
    get:
      operationId: getArchiveReconciliation
      tags: [archive]
      summary: Compare owner-scoped local archive records with Plex inventory
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated reconciliation report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationReport"
...
components:
...
  schemas:
...
    ReviewItemKind:
      type: string
      enum: [acquisition_recommendation, naming_proposal, archive_finding, operation_approval]
    ReviewItemState:
      type: string
      enum: [pending, approved, rejected, deferred, reopened]
...
    RecommendationConfidence:
      type: string
      enum: [high, medium, low, unknown]
    RecommendationPriority:
      type: string
      enum: [critical, high, medium, low]
    AcquisitionRecommendation:
      type: object
      required: [id, recommendationKey, mediaType, title, year, externalId, target, evidence, preferredQuality, destination, route, blockers, recommendedAction, confidence, priority, status, reviewItemId, acquisitionJobId, evidenceHash, generatedAt, updatedAt]
      properties:
        id: { type: number }
        recommendationKey: { type: string }
        mediaType: { type: string }
        title: { type: string }
        year: { type: ["number", "null"] }
        externalId: { type: ["string", "null"] }
        target:
          type: object
          additionalProperties: true
        evidence:
          type: object
          additionalProperties: true
        preferredQuality:
          type: object
          additionalProperties: true
        destination:
          type: object
          additionalProperties: true
        route:
          type: object
          additionalProperties: true
        blockers:
          type: array
          items: { type: string }
        recommendedAction: { type: string }
        confidence:
          $ref: "#/components/schemas/RecommendationConfidence"
        priority:
          $ref: "#/components/schemas/RecommendationPriority"
        status: { type: string }
        reviewItemId: { type: ["number", "null"] }
        acquisitionJobId: { type: ["number", "null"] }
        evidenceHash: { type: string }
        generatedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string